roxygen2md and split scpdsi.cpp
