+++ .gitignore
@@ -2,6 +2,8 @@
 .Rhistory
 .RData
 .Ruserdata
-src/*.o
-src/*.so
+
+*.o
+*.so
 src/*.dll
+inst/
