+++ DESCRIPTION
@@ -31,5 +31,6 @@ Imports:
     Rcpp (>= 0.12.0)
 URL: https://github.com/Sibada/scPDSI
 BugReports: https://github.com/Sibada/scPDSI/issues
-RoxygenNote: 5.0.1
+RoxygenNote: 6.1.1
 LinkingTo: Rcpp
+Roxygen: list(markdown = TRUE)
