+++ NAMESPACE
@@ -2,14 +2,23 @@
 
 S3method(plot,pdsi)
 export(pdsi)
+export(read_pdsi_input)
+export(read_pdsi_ouput)
+export(tidy_pdsi_df)
+import(magrittr)
 importFrom(Rcpp,sourceCpp)
+importFrom(data.table,fread)
 importFrom(graphics,abline)
 importFrom(graphics,lines)
 importFrom(graphics,plot)
 importFrom(graphics,points)
 importFrom(graphics,polygon)
+importFrom(lubridate,make_date)
+importFrom(purrr,is_empty)
+importFrom(purrr,map)
 importFrom(stats,end)
 importFrom(stats,frequency)
 importFrom(stats,start)
 importFrom(stats,ts)
+importFrom(zoo,zoo)
 useDynLib(scPDSI)
