+++ scPDSI.Rproj
@@ -6,8 +6,8 @@ AlwaysSaveHistory: Default
 
 EnableCodeIndexing: Yes
 UseSpacesForTab: Yes
-NumSpacesForTab: 2
-Encoding: CP936
+NumSpacesForTab: 4
+Encoding: UTF-8
 
 RnwWeave: Sweave
 LaTeX: pdfLaTeX
